[Bug] Pouvoir supprimer un compte utilisateur sans organisation depuis l’administration (#1403)
* Fixed bug that prevented the deletion of a user with no organization

* Trying to avoid false test errors

+++ src/accounts/signals.py
@@ -26,26 +26,27 @@ def manage_user_content_before_deletion(sender, instance, **kwargs):
     user = instance
     user_org = user.beneficiary_organization
 
-    # PROJECTS: Reattributed to another user belonging to the user's organization if such a user
-    # exists. Otherwise, the projects are deleted.
-    other_org_member = user_org.beneficiaries.exclude(id=user.id).first()
+    if user_org:
+        # PROJECTS: Reattributed to another user belonging to the user's organization if such a
+        # user exists. Otherwise, the projects are deleted.
+        other_org_member = user_org.beneficiaries.exclude(id=user.id).first()
 
-    projects = user_org.project_set.filter(author=user)
-    if other_org_member:
-        for project in projects:
-            project.author.add(other_org_member)
-            project.author.remove(user)
-            project.save()
-    else:
-        projects.delete()
+        projects = user_org.project_set.filter(author=user)
+        if other_org_member:
+            for project in projects:
+                project.author.add(other_org_member)
+                project.author.remove(user)
+                project.save()
+        else:
+            projects.delete()
 
-    # INVITATIONS (to the organization) : The form allows the user to reattribute them to another
-    #  user. Otherwise, they are deleted.
-    invited_users = User.objects.filter(invitation_author_id=user)
-    cancel_invitations(invited_users)
+        # INVITATIONS (to the organization) : The form allows the user to reattribute them to
+        # another user. Otherwise, they are deleted.
+        invited_users = User.objects.filter(invitation_author_id=user)
+        cancel_invitations(invited_users)
 
-    # ORGANIZATION: if the user is alone in their organization, it must be deleted too.
-    if not other_org_member:
-        user.beneficiary_organization = None
-        user.save()
-        user_org.delete()
+        # ORGANIZATION: if the user is alone in their organization, it must be deleted too.
+        if not other_org_member:
+            user.beneficiary_organization = None
+            user.save()
+            user_org.delete()
